feat: create phosphor-geometry + phosphor-engine-types, break upward coupling
Two new LGPL libraries that fix the #1 architectural problem: phosphor-zones
and phosphor-tiles no longer depend on phosphor-engine-api.

phosphor-geometry (SHARED, zero phosphor deps):
  Pure geometry math — snapToRect, coordinate transforms, overlap resolution,
  min-size enforcement, rectToJson. Depends only on Qt6::Core + Qt6::Gui.

phosphor-engine-types (INTERFACE, header-only):
  IPlacementState, NavigationContext, EngineTypes (TilingStateKey, SnapResult,
  ResnapEntry, PendingRestore, etc.). Depends only on Qt6::Core.

Dependency changes:
  phosphor-zones: engine-api → engine-types + geometry (no more engine-api)
  phosphor-tiles: engine-api → engine-types (no more engine-api)
  phosphor-engine-api: now depends on engine-types + geometry (transitive)

Engine-api retains forwarding headers so existing #include paths still work.
serializeZoneAssignments stays in engine-api (depends on EngineTypes).

Target dependency graph is now clean:
  engine-types ← zones, tiles (domain depends on types, not engine contract)
  geometry     ← zones (zone-aware geometry functions)
  engine-api   ← engine-types, geometry (engine contract layer)
  No zones ↔ tiles dependency. No domain → engine-contract dependency.

All 131 tests pass. Zero behavioral changes.
