+++ libs/phosphor-engine-api/include/PhosphorEngineApi/EngineTypes.h
@@ -3,92 +3,4 @@
 
 #pragma once
 
-#include <phosphorengineapi_export.h>
-
-#include <QHashFunctions>
-#include <QList>
-#include <QRect>
-#include <QSet>
-#include <QString>
-#include <QStringList>
-
-namespace PhosphorEngineApi {
-
-struct PHOSPHORENGINEAPI_EXPORT TilingStateKey
-{
-    QString screenId;
-    int desktop = 1;
-    QString activity;
-
-    bool operator==(const TilingStateKey& other) const
-    {
-        return screenId == other.screenId && desktop == other.desktop && activity == other.activity;
-    }
-};
-
-inline size_t qHash(const TilingStateKey& key, size_t seed = 0)
-{
-    return qHashMulti(seed, key.screenId, key.desktop, key.activity);
-}
-
-enum class SnapIntent {
-    UserInitiated,
-    AutoRestored,
-};
-
-struct PHOSPHORENGINEAPI_EXPORT ResnapEntry
-{
-    QString windowId;
-    int zonePosition;
-    QList<int> allZonePositions;
-    QString screenId;
-    int virtualDesktop = 0;
-};
-
-struct PHOSPHORENGINEAPI_EXPORT PendingRestore
-{
-    QStringList zoneIds;
-    QString screenId;
-    int virtualDesktop = 0;
-    QString layoutId;
-    QList<int> zoneNumbers;
-};
-
-struct PHOSPHORENGINEAPI_EXPORT SnapResult
-{
-    bool shouldSnap = false;
-    QRect geometry;
-    QString zoneId;
-    QStringList zoneIds;
-    QString screenId;
-
-    bool isValid() const
-    {
-        return shouldSnap && geometry.isValid() && !zoneId.isEmpty();
-    }
-
-    static SnapResult noSnap()
-    {
-        return SnapResult{false, QRect(), QString(), QStringList(), QString()};
-    }
-};
-
-struct PHOSPHORENGINEAPI_EXPORT UnfloatResult
-{
-    bool found = false;
-    QStringList zoneIds;
-    QRect geometry;
-    QString screenId;
-};
-
-struct PHOSPHORENGINEAPI_EXPORT ZoneAssignmentEntry
-{
-    QString windowId{};
-    QString sourceZoneId{};
-    QString targetZoneId{};
-    QStringList targetZoneIds{};
-    QRect targetGeometry{};
-    QString targetScreenId{};
-};
-
-} // namespace PhosphorEngineApi
+#include <PhosphorEngineTypes/EngineTypes.h>
