+++ libs/phosphor-engine-api/include/PhosphorEngineApi/GeometryUtils.h
@@ -4,31 +4,20 @@
 #pragma once
 
 #include <PhosphorEngineApi/EngineTypes.h>
+#include <PhosphorGeometry/GeometryUtils.h>
 #include <phosphorengineapi_export.h>
 
-#include <QRect>
-#include <QRectF>
-#include <QSize>
 #include <QString>
 #include <QVector>
-
-class QScreen;
 
 namespace PhosphorEngineApi {
 namespace GeometryUtils {
 
-PHOSPHORENGINEAPI_EXPORT QRectF availableAreaToOverlayCoordinates(const QRectF& geometry, QScreen* screen);
-PHOSPHORENGINEAPI_EXPORT QRectF availableAreaToOverlayCoordinates(const QRectF& geometry, const QRect& overlayGeometry);
-
-PHOSPHORENGINEAPI_EXPORT QRect snapToRect(const QRectF& rf);
-
-PHOSPHORENGINEAPI_EXPORT void enforceWindowMinSizes(QVector<QRect>& zones, const QVector<QSize>& minSizes,
-                                                    int gapThreshold, int innerGap = 0);
-
-PHOSPHORENGINEAPI_EXPORT void removeZoneOverlaps(QVector<QRect>& zones, const QVector<QSize>& minSizes = {},
-                                                 int innerGap = 0);
-
-PHOSPHORENGINEAPI_EXPORT QString rectToJson(const QRect& rect);
+using PhosphorGeometry::availableAreaToOverlayCoordinates;
+using PhosphorGeometry::enforceWindowMinSizes;
+using PhosphorGeometry::rectToJson;
+using PhosphorGeometry::removeZoneOverlaps;
+using PhosphorGeometry::snapToRect;
 
 PHOSPHORENGINEAPI_EXPORT QString serializeZoneAssignments(const QVector<ZoneAssignmentEntry>& entries);
