+++ libs/phosphor-engine-types/PhosphorEngineTypesConfig.cmake.in
@@ -1,0 +1,12 @@
+# SPDX-FileCopyrightText: 2026 fuddlesworth
+# SPDX-License-Identifier: LGPL-2.1-or-later
+
+@PACKAGE_INIT@
+
+include(CMakeFindDependencyMacro)
+
+find_dependency(Qt6 6.6 COMPONENTS Core)
+
+include("${CMAKE_CURRENT_LIST_DIR}/PhosphorEngineTypesTargets.cmake")
+
+check_required_components(PhosphorEngineTypes)
