+++ libs/phosphor-geometry/include/PhosphorGeometry/JsonKeys.h
@@ -1,0 +1,22 @@
+// SPDX-FileCopyrightText: 2026 fuddlesworth
+// SPDX-License-Identifier: LGPL-2.1-or-later
+
+#pragma once
+
+#include <QLatin1String>
+
+namespace PhosphorGeometry {
+namespace JsonKeys {
+
+inline constexpr QLatin1String X{"x"};
+inline constexpr QLatin1String Y{"y"};
+inline constexpr QLatin1String Width{"width"};
+inline constexpr QLatin1String Height{"height"};
+
+inline constexpr QLatin1String WindowId{"windowId"};
+inline constexpr QLatin1String SourceZoneId{"sourceZoneId"};
+inline constexpr QLatin1String TargetZoneId{"targetZoneId"};
+inline constexpr QLatin1String TargetZoneIds{"targetZoneIds"};
+
+} // namespace JsonKeys
+} // namespace PhosphorGeometry
